backend/neurofin_backend/agent/requirements.txt
flask==2.3.2
requests==2.31.0
openai==1.0.0
python-dotenv==1.0.0
weaviate-client==4.18.0
langgraph>=1.0.3
faiss-cpu>=1.7.4
tqdm
pymongo
groq





